보스1 Jupiter, 체력시스템 구현
보스1 기본 코드를 작성하고
체력 시스템을 구현하고 기본 UI를 추가하였습니다.

+++ Assets/Script/MoveObject.cs
@@ -5,9 +5,9 @@ using UnityEngine;
 
 public class MoveObject : MonoBehaviour
 {
-    protected float _HP;
+    protected int _HP;
     protected float _speed;
-    protected float _damage;
+    protected int _damage;
     protected float _width;
     protected float _height;
